Update Tables.html
Please double-check that I am lining up the columns correctly, but insert for a singly linked list should be O(N) time. The reason is that inserting at the end of a list requires traversal throughout the entire list, since a singly linked list lacks the backwards pointer of a doubly linked list.